Adapt all tests to new integer parsing
The tests now expect the correct types specified in the tables schema

+++ sql/parse/parse_test.go
@@ -181,7 +181,7 @@ var fixtures = map[string]sql.Node{
 				plan.NewFilter(
 					expression.NewEquals(
 						expression.NewUnresolvedColumn("qux"),
-						expression.NewLiteral(int64(1), sql.Int64),
+						expression.NewLiteral(int8(1), sql.Int8),
 					),
 					plan.NewUnresolvedTable("foo", ""),
 				),
@@ -258,7 +258,7 @@ var fixtures = map[string]sql.Node{
 		plan.NewUnresolvedTable("t1", ""),
 		plan.NewValues([][]sql.Expression{{
 			expression.NewLiteral("a", sql.Text),
-			expression.NewLiteral(int64(1), sql.Int64),
+			expression.NewLiteral(int8(1), sql.Int8),
 		}}),
 		false,
 		[]string{"col1", "col2"},
@@ -267,7 +267,7 @@ var fixtures = map[string]sql.Node{
 		plan.NewUnresolvedTable("t1", ""),
 		plan.NewValues([][]sql.Expression{{
 			expression.NewLiteral("a", sql.Text),
-			expression.NewLiteral(int64(1), sql.Int64),
+			expression.NewLiteral(int8(1), sql.Int8),
 		}}),
 		true,
 		[]string{"col1", "col2"},
@@ -360,7 +360,7 @@ var fixtures = map[string]sql.Node{
 		plan.NewFilter(
 			expression.NewEquals(
 				expression.NewUnresolvedColumn("a"),
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 			),
 			plan.NewUnresolvedTable("foo", ""),
 		),
@@ -432,9 +432,9 @@ var fixtures = map[string]sql.Node{
 		plan.NewFilter(
 			expression.NewNot(
 				expression.NewBetween(
-					expression.NewLiteral(int64(1), sql.Int64),
-					expression.NewLiteral(int64(2), sql.Int64),
-					expression.NewLiteral(int64(5), sql.Int64),
+					expression.NewLiteral(int8(1), sql.Int8),
+					expression.NewLiteral(int8(2), sql.Int8),
+					expression.NewLiteral(int8(5), sql.Int8),
 				),
 			),
 			plan.NewUnresolvedTable("foo", ""),
@@ -444,16 +444,16 @@ var fixtures = map[string]sql.Node{
 		[]sql.Expression{expression.NewStar()},
 		plan.NewFilter(
 			expression.NewBetween(
-				expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewLiteral(int64(2), sql.Int64),
-				expression.NewLiteral(int64(5), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewLiteral(int8(2), sql.Int8),
+				expression.NewLiteral(int8(5), sql.Int8),
 			),
 			plan.NewUnresolvedTable("foo", ""),
 		),
 	),
 	`SELECT 0x01AF`: plan.NewProject(
 		[]sql.Expression{
-			expression.NewLiteral(int64(431), sql.Int64),
+			expression.NewLiteral(int16(431), sql.Int16),
 		},
 		plan.NewUnresolvedTable("dual", ""),
 	),
@@ -470,12 +470,12 @@ var fixtures = map[string]sql.Node{
 				"somefunc",
 				false,
 				expression.NewTuple(
-					expression.NewLiteral(int64(1), sql.Int64),
-					expression.NewLiteral(int64(2), sql.Int64),
+					expression.NewLiteral(int8(1), sql.Int8),
+					expression.NewLiteral(int8(2), sql.Int8),
 				),
 				expression.NewTuple(
-					expression.NewLiteral(int64(3), sql.Int64),
-					expression.NewLiteral(int64(4), sql.Int64),
+					expression.NewLiteral(int8(3), sql.Int8),
+					expression.NewLiteral(int8(4), sql.Int8),
 				),
 			),
 			plan.NewUnresolvedTable("b", ""),
@@ -486,7 +486,7 @@ var fixtures = map[string]sql.Node{
 		plan.NewFilter(
 			expression.NewEquals(
 				expression.NewLiteral(":foo_id", sql.Text),
-				expression.NewLiteral(int64(2), sql.Int64),
+				expression.NewLiteral(int8(2), sql.Int8),
 			),
 			plan.NewUnresolvedTable("foo", ""),
 		),
@@ -510,13 +510,13 @@ var fixtures = map[string]sql.Node{
 	),
 	`SELECT CAST(-3 AS UNSIGNED) FROM foo`: plan.NewProject(
 		[]sql.Expression{
-			expression.NewConvert(expression.NewLiteral(int64(-3), sql.Int64), expression.ConvertToUnsigned),
+			expression.NewConvert(expression.NewLiteral(int8(-3), sql.Int8), expression.ConvertToUnsigned),
 		},
 		plan.NewUnresolvedTable("foo", ""),
 	),
 	`SELECT 2 = 2 FROM foo`: plan.NewProject(
 		[]sql.Expression{
-			expression.NewEquals(expression.NewLiteral(int64(2), sql.Int64), expression.NewLiteral(int64(2), sql.Int64)),
+			expression.NewEquals(expression.NewLiteral(int8(2), sql.Int8), expression.NewLiteral(int8(2), sql.Int8)),
 		},
 		plan.NewUnresolvedTable("foo", ""),
 	),
@@ -560,10 +560,10 @@ var fixtures = map[string]sql.Node{
 		[]sql.Expression{expression.NewStar()},
 		plan.NewFilter(
 			expression.NewIn(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				expression.NewTuple(
 					expression.NewLiteral("1", sql.Text),
-					expression.NewLiteral(int64(2), sql.Int64),
+					expression.NewLiteral(int8(2), sql.Int8),
 				),
 			),
 			plan.NewUnresolvedTable("foo", ""),
@@ -573,10 +573,10 @@ var fixtures = map[string]sql.Node{
 		[]sql.Expression{expression.NewStar()},
 		plan.NewFilter(
 			expression.NewNotIn(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				expression.NewTuple(
 					expression.NewLiteral("1", sql.Text),
-					expression.NewLiteral(int64(2), sql.Int64),
+					expression.NewLiteral(int8(2), sql.Int8),
 				),
 			),
 			plan.NewUnresolvedTable("foo", ""),
@@ -585,12 +585,12 @@ var fixtures = map[string]sql.Node{
 	`SELECT a, b FROM t ORDER BY 2, 1`: plan.NewSort(
 		[]plan.SortField{
 			{
-				Column:       expression.NewLiteral(int64(2), sql.Int64),
+				Column:       expression.NewLiteral(int8(2), sql.Int8),
 				Order:        plan.Ascending,
 				NullOrdering: plan.NullsFirst,
 			},
 			{
-				Column:       expression.NewLiteral(int64(1), sql.Int64),
+				Column:       expression.NewLiteral(int8(1), sql.Int8),
 				Order:        plan.Ascending,
 				NullOrdering: plan.NullsFirst,
 			},
@@ -605,22 +605,22 @@ var fixtures = map[string]sql.Node{
 	),
 	`SELECT 1 + 1;`: plan.NewProject(
 		[]sql.Expression{
-			expression.NewPlus(expression.NewLiteral(int64(1), sql.Int64), expression.NewLiteral(int64(1), sql.Int64)),
+			expression.NewPlus(expression.NewLiteral(int8(1), sql.Int8), expression.NewLiteral(int8(1), sql.Int8)),
 		},
 		plan.NewUnresolvedTable("dual", ""),
 	),
 	`SELECT 1 * (2 + 1);`: plan.NewProject(
 		[]sql.Expression{
-			expression.NewMult(expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewPlus(expression.NewLiteral(int64(2), sql.Int64), expression.NewLiteral(int64(1), sql.Int64))),
+			expression.NewMult(expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewPlus(expression.NewLiteral(int8(2), sql.Int8), expression.NewLiteral(int8(1), sql.Int8))),
 		},
 		plan.NewUnresolvedTable("dual", ""),
 	),
 	`SELECT (0 - 1) * (1 | 1);`: plan.NewProject(
 		[]sql.Expression{
 			expression.NewMult(
-				expression.NewMinus(expression.NewLiteral(int64(0), sql.Int64), expression.NewLiteral(int64(1), sql.Int64)),
-				expression.NewBitOr(expression.NewLiteral(int64(1), sql.Int64), expression.NewLiteral(int64(1), sql.Int64)),
+				expression.NewMinus(expression.NewLiteral(int8(0), sql.Int8), expression.NewLiteral(int8(1), sql.Int8)),
+				expression.NewBitOr(expression.NewLiteral(int8(1), sql.Int8), expression.NewLiteral(int8(1), sql.Int8)),
 			),
 		},
 		plan.NewUnresolvedTable("dual", ""),
@@ -628,8 +628,8 @@ var fixtures = map[string]sql.Node{
 	`SELECT (1 << 3) % (2 div 1);`: plan.NewProject(
 		[]sql.Expression{
 			expression.NewMod(
-				expression.NewShiftLeft(expression.NewLiteral(int64(1), sql.Int64), expression.NewLiteral(int64(3), sql.Int64)),
-				expression.NewIntDiv(expression.NewLiteral(int64(2), sql.Int64), expression.NewLiteral(int64(1), sql.Int64))),
+				expression.NewShiftLeft(expression.NewLiteral(int8(1), sql.Int8), expression.NewLiteral(int8(3), sql.Int8)),
+				expression.NewIntDiv(expression.NewLiteral(int8(2), sql.Int8), expression.NewLiteral(int8(1), sql.Int8))),
 		},
 		plan.NewUnresolvedTable("dual", ""),
 	),
@@ -645,7 +645,7 @@ var fixtures = map[string]sql.Node{
 	`SELECT '1.0' + 2;`: plan.NewProject(
 		[]sql.Expression{
 			expression.NewPlus(
-				expression.NewLiteral("1.0", sql.Text), expression.NewLiteral(int64(2), sql.Int64),
+				expression.NewLiteral("1.0", sql.Text), expression.NewLiteral(int8(2), sql.Int8),
 			),
 		},
 		plan.NewUnresolvedTable("dual", ""),
@@ -714,7 +714,7 @@ var fixtures = map[string]sql.Node{
 				expression.NewUnresolvedFunction(
 					"max", true, expression.NewUnresolvedColumn("i"),
 				),
-				expression.NewLiteral(int64(2), sql.Int64),
+				expression.NewLiteral(int8(2), sql.Int8),
 				"/",
 			),
 		},
@@ -742,7 +742,7 @@ var fixtures = map[string]sql.Node{
 	`SET autocommit=1, foo="bar"`: plan.NewSet(
 		plan.SetVariable{
 			Name:  "autocommit",
-			Value: expression.NewLiteral(int64(1), sql.Int64),
+			Value: expression.NewLiteral(int8(1), sql.Int8),
 		},
 		plan.SetVariable{
 			Name:  "foo",
@@ -752,7 +752,7 @@ var fixtures = map[string]sql.Node{
 	`SET @@session.autocommit=1, foo="bar"`: plan.NewSet(
 		plan.SetVariable{
 			Name:  "@@session.autocommit",
-			Value: expression.NewLiteral(int64(1), sql.Int64),
+			Value: expression.NewLiteral(int8(1), sql.Int8),
 		},
 		plan.SetVariable{
 			Name:  "foo",
@@ -818,11 +818,11 @@ var fixtures = map[string]sql.Node{
 	`SET SESSION NET_READ_TIMEOUT= 700, SESSION NET_WRITE_TIMEOUT= 700`: plan.NewSet(
 		plan.SetVariable{
 			Name:  "@@session.net_read_timeout",
-			Value: expression.NewLiteral(int64(700), sql.Int64),
+			Value: expression.NewLiteral(int16(700), sql.Int16),
 		},
 		plan.SetVariable{
 			Name:  "@@session.net_write_timeout",
-			Value: expression.NewLiteral(int64(700), sql.Int64),
+			Value: expression.NewLiteral(int16(700), sql.Int16),
 		},
 	),
 	`SET gtid_mode=DEFAULT`: plan.NewSet(
@@ -975,11 +975,11 @@ var fixtures = map[string]sql.Node{
 			expression.NewUnresolvedColumn("foo"),
 			[]expression.CaseBranch{
 				{
-					Cond:  expression.NewLiteral(int64(1), sql.Int64),
+					Cond:  expression.NewLiteral(int8(1), sql.Int8),
 					Value: expression.NewLiteral("foo", sql.Text),
 				},
 				{
-					Cond:  expression.NewLiteral(int64(2), sql.Int64),
+					Cond:  expression.NewLiteral(int8(2), sql.Int8),
 					Value: expression.NewLiteral("bar", sql.Text),
 				},
 			},
@@ -992,11 +992,11 @@ var fixtures = map[string]sql.Node{
 			expression.NewUnresolvedColumn("foo"),
 			[]expression.CaseBranch{
 				{
-					Cond:  expression.NewLiteral(int64(1), sql.Int64),
+					Cond:  expression.NewLiteral(int8(1), sql.Int8),
 					Value: expression.NewLiteral("foo", sql.Text),
 				},
 				{
-					Cond:  expression.NewLiteral(int64(2), sql.Int64),
+					Cond:  expression.NewLiteral(int8(2), sql.Int8),
 					Value: expression.NewLiteral("bar", sql.Text),
 				},
 			},
@@ -1011,14 +1011,14 @@ var fixtures = map[string]sql.Node{
 				{
 					Cond: expression.NewEquals(
 						expression.NewUnresolvedColumn("foo"),
-						expression.NewLiteral(int64(1), sql.Int64),
+						expression.NewLiteral(int8(1), sql.Int8),
 					),
 					Value: expression.NewLiteral("foo", sql.Text),
 				},
 				{
 					Cond: expression.NewEquals(
 						expression.NewUnresolvedColumn("foo"),
-						expression.NewLiteral(int64(2), sql.Int64),
+						expression.NewLiteral(int8(2), sql.Int8),
 					),
 					Value: expression.NewLiteral("bar", sql.Text),
 				},
@@ -1055,7 +1055,7 @@ var fixtures = map[string]sql.Node{
 		[]sql.Expression{expression.NewArithmetic(
 			expression.NewLiteral("2018-05-01", sql.Text),
 			expression.NewInterval(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				"DAY",
 			),
 			"+",
@@ -1066,7 +1066,7 @@ var fixtures = map[string]sql.Node{
 		[]sql.Expression{expression.NewArithmetic(
 			expression.NewLiteral("2018-05-01", sql.Text),
 			expression.NewInterval(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				"DAY",
 			),
 			"-",
@@ -1076,7 +1076,7 @@ var fixtures = map[string]sql.Node{
 	`SELECT INTERVAL 1 DAY + '2018-05-01'`: plan.NewProject(
 		[]sql.Expression{expression.NewArithmetic(
 			expression.NewInterval(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				"DAY",
 			),
 			expression.NewLiteral("2018-05-01", sql.Text),
@@ -1089,13 +1089,13 @@ var fixtures = map[string]sql.Node{
 			expression.NewArithmetic(
 				expression.NewLiteral("2018-05-01", sql.Text),
 				expression.NewInterval(
-					expression.NewLiteral(int64(1), sql.Int64),
+					expression.NewLiteral(int8(1), sql.Int8),
 					"DAY",
 				),
 				"+",
 			),
 			expression.NewInterval(
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
 				"DAY",
 			),
 			"+",
@@ -1105,7 +1105,7 @@ var fixtures = map[string]sql.Node{
 	`SELECT COUNT(*) FROM foo GROUP BY a HAVING COUNT(*) > 5`: plan.NewHaving(
 		expression.NewGreaterThan(
 			expression.NewUnresolvedFunction("count", true, expression.NewStar()),
-			expression.NewLiteral(int64(5), sql.Int64),
+			expression.NewLiteral(int8(5), sql.Int8),
 		),
 		plan.NewGroupBy(
 			[]sql.Expression{expression.NewUnresolvedFunction("count", true, expression.NewStar())},
@@ -1117,7 +1117,7 @@ var fixtures = map[string]sql.Node{
 		plan.NewHaving(
 			expression.NewGreaterThan(
 				expression.NewUnresolvedFunction("count", true, expression.NewStar()),
-				expression.NewLiteral(int64(5), sql.Int64),
+				expression.NewLiteral(int8(5), sql.Int8),
 			),
 			plan.NewGroupBy(
 				[]sql.Expression{expression.NewUnresolvedFunction("count", true, expression.NewStar())},
@@ -1132,8 +1132,8 @@ var fixtures = map[string]sql.Node{
 			plan.NewUnresolvedTable("foo", ""),
 			plan.NewUnresolvedTable("bar", ""),
 			expression.NewEquals(
-				expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewLiteral(int8(1), sql.Int8),
 			),
 		),
 	),
@@ -1143,8 +1143,8 @@ var fixtures = map[string]sql.Node{
 			plan.NewUnresolvedTable("foo", ""),
 			plan.NewUnresolvedTable("bar", ""),
 			expression.NewEquals(
-				expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewLiteral(int8(1), sql.Int8),
 			),
 		),
 	),
@@ -1154,8 +1154,8 @@ var fixtures = map[string]sql.Node{
 			plan.NewUnresolvedTable("foo", ""),
 			plan.NewUnresolvedTable("bar", ""),
 			expression.NewEquals(
-				expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewLiteral(int8(1), sql.Int8),
 			),
 		),
 	),
@@ -1165,8 +1165,8 @@ var fixtures = map[string]sql.Node{
 			plan.NewUnresolvedTable("foo", ""),
 			plan.NewUnresolvedTable("bar", ""),
 			expression.NewEquals(
-				expression.NewLiteral(int64(1), sql.Int64),
-				expression.NewLiteral(int64(1), sql.Int64),
+				expression.NewLiteral(int8(1), sql.Int8),
+				expression.NewLiteral(int8(1), sql.Int8),
 			),
 		),
 	),
